Refactor onboarding5 to use functional component
Why:
We would like to introduce some information to the component via a
context. It is best practice to use hooks and so we would like to use
the useContext hook for this. Currently the onboarding5 component is
build with the class syntax keeping us from doing this.

This commit:
Refactors the Onbarding5 component to use the functional style. No new
functionality has been added.