Feature/bte tp onboard (#863)
* Update the onboard flow for Bluetooth

Why:
When using the BTE tracing strategy, the onboarding flow needs to change in order to
ask for bluetooth permissions. At this time, the bluetooth permissions are not properly
wired up. In the future, this will be fixed or the Exposure Notification logic from
Google/Apple will handle it.

This commit:
Adds stand-in copy for the time being to the bluetooth onboarding flow. Add bluetooth
permissions to ios. Add bluetooth logic to Onboard5.js.

Fix for info.plist relative paths issue.
Fix package.json run-ios aliases - updating the scheme