Move news components to feature folder
Why:
Currently the components folder has a NewsItem Component which is only
being used in the NewsScreen.

This commit:
Moves the news item and the news screen to a user facing feature folder
for better organization as the news item component is not intended to be
used generically.

Additionally we removed the link to the news feed in
the settings screen as the News Screen itself is not included in any
existing navigator and will cause a crash if a user taps it.

It might makes sense to remove the news screen altogether if it is no
longer a part of the product feature set, but we are not sure if it is
or not at this point.